# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Investment Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: Get all assets for current user
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [equity_fund, debt_fund, metal, cash]
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetBody"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetail"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetBody"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /assets/{id}/valuations:
    get:
      operationId: getAssetValuations
      tags: [assets]
      summary: Get valuation history for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Valuations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Valuation"
    post:
      operationId: addAssetValuation
      tags: [assets]
      summary: Add a valuation snapshot for an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddValuationBody"
      responses:
        "201":
          description: Valuation added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Valuation"
  /assets/{id}/valuations/{valuationId}:
    patch:
      operationId: updateAssetValuation
      tags: [assets]
      summary: Update a valuation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: valuationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddValuationBody"
      responses:
        "200":
          description: Valuation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Valuation"
    delete:
      operationId: deleteAssetValuation
      tags: [assets]
      summary: Delete a valuation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: valuationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Transactions
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: Get all cash accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a cash account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update a cash account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete a cash account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [equity_fund, debt_fund, metal, cash]
        subCategory:
          type: ["string", "null"]
        currentValue:
          type: number
        investedValue:
          type: number
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        profitLoss:
          type: number
        profitLossPercent:
          type: number
        targetPercent:
          type: number
        actualPercent:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - currentValue
        - investedValue
        - profitLoss
        - profitLossPercent
        - targetPercent
        - actualPercent
        - currency
        - isActive
        - createdAt
        - updatedAt
    AssetDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        currentValue:
          type: number
        investedValue:
          type: number
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        profitLoss:
          type: number
        profitLossPercent:
          type: number
        targetPercent:
          type: number
        actualPercent:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        valuations:
          type: array
          items:
            $ref: "#/components/schemas/Valuation"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - id
        - name
        - category
        - currentValue
        - investedValue
        - profitLoss
        - profitLossPercent
        - targetPercent
        - actualPercent
        - currency
        - isActive
        - createdAt
        - updatedAt
        - valuations
        - transactions
    CreateAssetBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [equity_fund, debt_fund, metal, cash]
        subCategory:
          type: ["string", "null"]
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        investedValue:
          type: number
        targetPercent:
          type: number
        currency:
          type: string
      required:
        - name
        - category
        - investedValue
        - targetPercent
        - currency
    UpdateAssetBody:
      type: object
      properties:
        name:
          type: string
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        investedValue:
          type: number
        targetPercent:
          type: number
        isActive:
          type: boolean
    Valuation:
      type: object
      properties:
        id:
          type: integer
        assetId:
          type: integer
        value:
          type: number
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        date:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - assetId
        - value
        - date
        - createdAt
    AddValuationBody:
      type: object
      properties:
        value:
          type: number
        units:
          type: ["number", "null"]
        nav:
          type: ["number", "null"]
        pricePerUnit:
          type: ["number", "null"]
        date:
          type: string
        note:
          type: ["string", "null"]
      required:
        - value
        - date

    # Transactions
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, transfer, invest, redeem, tag_allocation, sip]
        amount:
          type: number
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        sourceAccountId:
          type: ["integer", "null"]
        destinationAccountId:
          type: ["integer", "null"]
        date:
          type: string
        tag:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - date
        - createdAt
...
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [bank, money_market, cash]
        tag:
          type: string
          enum: [emergency, opportunity, free]
        balance:
          type: number
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - tag
        - balance
        - currency
        - isActive
        - createdAt
        - updatedAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [bank, money_market, cash]
        tag:
          type: string
          enum: [emergency, opportunity, free]
        balance:
          type: number
        currency:
          type: string
      required:
        - name
        - type
        - tag
        - balance
        - currency
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        tag:
          type: string
        balance:
          type: number
        isActive:
          type: boolean
...